Code organization
World, meet Librari Anna. 34-th century robotic communicator of knowledge and stories.

Moved _The AquariYumYum Disc Horse_ to a book. The first of its kind. Cherish its participation in the joys of Surce.

May the .* move to Surce!

-- src/librarianna.jsx
